backend/requirements.txt
fastapi
uvicorn
langchain-google-genai
pydantic
python-dotenv
httpx
requests
supabase